+++ Runtime/Interpreter/TokenType.cs
@@ -7,7 +7,7 @@ using System.Threading.Tasks;
 namespace RealityProgrammer.CSStandard.Interpreter {
     public enum TokenType {
         LeftParenthesis, RightParenthesis, Dot, Comma, Greater, Less, Equal, Bang,
-        Plus, Minus, Star, Slash, Question,
+        Plus, Minus, Star, Slash, Question, Percentage,
         
         BitwiseAnd, BitwiseOr, BitwiseXor, BitwiseLeftShift, BitwiseRightShift, BitwiseComplement,
         BitwiseAndEqual, BitwiseOrEqual, BitwiseXorEqual, BitwiseLeftShiftEqual, BitwiseRightShiftEqual, BitwiseComplementEqual,