+++ CHANGELOG.md
@@ -3,6 +3,15 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/)  
 
+## [1.2.0] - 2020-10-01
+### Added
+- Collection of thing about Shader Property, include runtime modifier (for both UI and Renderer)
+- Storage of runtime created Materials.
+- Addition algorithms for AlgorithmUtilities.
+
+### Changed
+- Increased performance of EquationCalculatorInterpreter
+
 ## [1.1.0] - 2020-09-30
 ### Added
 - Equation Calculator.